+++ CHANGELOG.md
@@ -4,6 +4,11 @@
   by `Predicate.coerce(...)`, with `x = 12 and y in ('foo','bar')`
   semantics
 
+* Adds `Predicate.to_sequel` that compiles predicates to Sequel
+  expressions. Note that Sequel is not a dependency of Predicate and
+  must be required by the user. `Predicate.to_sequel` is only available
+  if `require 'predicate/sequel'` is done first.
+
 # 1.0.0 / 2018-03-03
 
 Predicate 1.0.0, extracted & refactored from alf-core.